+++ tests/Integration/inc/Engine/Admin/PerformanceMonitoring/Subscriber/RenderLicenseBannerSectionTest.php
@@ -1,4 +1,5 @@
 <?php
+declare( strict_types=1 );
 
 namespace WP_Rocket\Tests\Integration\inc\Engine\Admin\PerformanceMonitoring\Subscriber;
 
@@ -9,7 +10,7 @@ use WP_Rocket\Tests\Integration\TestCase;
  *
  * @group PerformanceMonitoring
  */
-class TestRenderLicenseBannerSection extends TestCase {
+class RenderLicenseBannerSectionTest extends TestCase {
 
     public function set_up() {
         parent::set_up();
