+++ tests/Integration/inc/Engine/Admin/PerformanceMonitoring/Subscriber/ResetPageTest.php
@@ -1,4 +1,5 @@
 <?php
+declare( strict_types=1 );
 
 namespace WP_Rocket\Tests\Integration\inc\Engine\Admin\PerformanceMonitoring\Subscriber;
 
@@ -6,15 +7,16 @@ use WP_Rocket\Tests\Integration\DBTrait;
 use WP_Rocket\Tests\Integration\AjaxTestCase;
 
 /**
- * Test class covering WP_Rocket\Engine\Admin\PerformanceMonitoring\Subscriber::add_new_page
+ * Test class covering WP_Rocket\Engine\Admin\PerformanceMonitoring\Subscriber::reset_page
  *
  * @group PerformanceMonitoring
  * @group AdminOnly
  */
-class Test_AddNewPage extends AjaxTestCase {
+class ResetPageTest extends AjaxTestCase {
 	use DBTrait;
 
 	private $hook_fired = false;
+	private $hook_fired_id = null;
 
 	public static function set_up_before_class() {
 		parent::set_up_before_class();
@@ -39,12 +41,13 @@ class Test_AddNewPage extends AjaxTestCase {
 		add_filter( 'rocket_performance_monitoring_enabled', '__return_true' );
 
 		// Set the AJAX action
-		$this->action = 'rocket_pm_add_new_page';
+		$this->action = 'rocket_pm_reset_page';
 
-		// Add a hook to capture when rocket_pm_job_added is fired
-		add_action( 'rocket_pm_job_added', [ $this, 'capture_hook_fired' ] );
+		// Add a hook to capture when rocket_pm_job_retest is fired
+		add_action( 'rocket_pm_job_retest', [ $this, 'capture_hook_fired' ] );
 
 		$this->hook_fired = false;
+		$this->hook_fired_id = null;
 	}
 
 	public function tear_down() {
@@ -54,14 +57,8 @@ class Test_AddNewPage extends AjaxTestCase {
 		// Remove Performance Monitoring enabled filter
 		remove_filter( 'rocket_performance_monitoring_enabled', '__return_true' );
 
-		// Remove URL limit filter if set
-		remove_filter( 'wpr_pm_allow_add_page', '__return_false' );
-
 		// Remove our test hook
-		remove_action( 'rocket_pm_job_added', [ $this, 'capture_hook_fired' ] );
-
-		// Remove mock HTTP filter
-		remove_filter( 'pre_http_request', [ $this, 'mock_http_request' ] );
+		remove_action( 'rocket_pm_job_retest', [ $this, 'capture_hook_fired' ] );
 
 		parent::tear_down();
 	}
@@ -75,31 +72,24 @@ class Test_AddNewPage extends AjaxTestCase {
 		$this->executeAjaxCall();
 
 		$this->assertResponse( $expected );
-
-		$this->cleanUpTest( $config );
 	}
 
 	private function setUpTest( $config ) {
 		// Set up the nonce
 		$_POST['nonce'] = \wp_create_nonce( 'rocket-ajax' );
 
+		// Set up database entries if provided
+		if ( isset( $config['database_entries'] ) ) {
+			foreach ( $config['database_entries'] as $entry ) {
+				self::addPerformanceMonitoring( $entry );
+			}
+		}
+
 		// Set up POST data if provided
 		if ( isset( $config['post_data'] ) ) {
 			foreach ( $config['post_data'] as $key => $value ) {
 				$_POST[ $key ] = $value;
 			}
-		}
-
-		// Set up filters if provided
-		if ( isset( $config['filters'] ) ) {
-			foreach ( $config['filters'] as $filter => $callback ) {
-				add_filter( $filter, $callback );
-			}
-		}
-
-		// Mock HTTP requests if needed for URL validation
-		if ( isset( $config['mock_http'] ) && $config['mock_http'] ) {
-			add_filter( 'pre_http_request', [ $this, 'mock_http_request' ], 10, 3 );
 		}
 	}
 
@@ -128,23 +118,29 @@ class Test_AddNewPage extends AjaxTestCase {
 
 	private function assertSuccessResponse( $response, $expected ) {
 		$this->assertTrue( $response['success'] );
-		
-		// Check if database entry was created
-		if ( isset( $expected['database_entries'] ) && $expected['database_entries'] > 0 ) {
-			$container = apply_filters( 'rocket_container', null );
-			$pm_query = $container->get( 'pm_query' );
-			$items = $pm_query->query( [] );
-			$this->assertSame( $expected['database_entries'], count( $items ) );
-		}
 
 		// Check if hook was fired
 		if ( isset( $expected['hook_fired'] ) && $expected['hook_fired'] ) {
 			$this->assertTrue( $this->hook_fired );
+
+			if ( isset( $expected['hook_fired_id'] ) ) {
+				$this->assertSame( $expected['hook_fired_id'], $this->hook_fired_id );
+			}
 		}
 
 		// Check response data if provided
 		if ( isset( $expected['response_data'] ) ) {
 			foreach ( $expected['response_data'] as $key => $value ) {
+				$this->assertArrayHasKey( $key, $response['data'] );
+				if ( $value !== null ) {
+					$this->assertSame( $value, $response['data'][ $key ] );
+				}
+			}
+		}
+
+		// Check that response contains expected keys
+		if ( isset( $expected['response_keys'] ) ) {
+			foreach ( $expected['response_keys'] as $key ) {
 				$this->assertArrayHasKey( $key, $response['data'] );
 			}
 		}
@@ -152,56 +148,25 @@ class Test_AddNewPage extends AjaxTestCase {
 
 	private function assertErrorResponse( $response, $expected ) {
 		$this->assertFalse( $response['success'] );
-		
+
 		// Check error message if provided
 		if ( isset( $expected['error_message'] ) ) {
 			$this->assertStringContainsString( $expected['error_message'], $response['data']['message'] );
 		}
-	}
 
-	private function cleanUpTest( $config ) {
-		// Clean up filters
-		if ( isset( $config['filters'] ) ) {
-			foreach ( $config['filters'] as $filter => $callback ) {
-				remove_filter( $filter, $callback );
-			}
+		// Check if hook was NOT fired for error cases
+		if ( isset( $expected['hook_fired'] ) && ! $expected['hook_fired'] ) {
+			$this->assertFalse( $this->hook_fired );
 		}
 	}
 
 	/**
-	 * Callback to capture when rocket_pm_job_added hook is fired.
+	 * Callback to capture when rocket_pm_job_retest hook is fired.
 	 *
-	 * @param string $url The URL that was added for monitoring.
+	 * @param int $id The database row ID of the reset job.
 	 */
-	public function capture_hook_fired( $url ) {
+	public function capture_hook_fired( $id ) {
 		$this->hook_fired = true;
-	}
-
-	/**
-	 * Mock HTTP requests for URL validation.
-	 *
-	 * @param false|array|WP_Error $preempt A preemptive return value of an HTTP request.
-	 * @param array                $args HTTP request arguments.
-	 * @param string               $url The request URL.
-	 * @return array|false
-	 */
-	public function mock_http_request( $preempt, $args, $url ) {
-		// Mock successful response for URLs on the test domain (example.org)
-		if ( strpos( $url, 'http://example.org' ) === 0 ) {
-			return [
-				'response' => [
-					'code' => 200,
-				],
-				'body' => '<html><head><title>Test Page Title</title></head><body>Test content</body></html>',
-			];
-		}
-
-		// Mock 404 for invalid URLs
-		return [
-			'response' => [
-				'code' => 404,
-			],
-			'body' => 'Not found',
-		];
+		$this->hook_fired_id = $id;
 	}
 }
